# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/core/district": {
      "post": {
        "description": "เพิ่มข้อมูลตำบล",
        "operationId": "DistrictController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDistrictDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operation successful.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Operation successful.",
                    "error": 0,
                    "data": {
                      "id": 1,
                      "name": "ยโสธร"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": [
                      "Invalid input"
                    ],
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "ต้องเข้าสู่ระบบก่อน",
                    "error": "Unauthorized",
                    "statusCode": 401
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "คุณไม่ได้รับอนุญาติให้กระทำสิ่งนี้",
                    "error": "Forbidden",
                    "statusCode": 401
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Resource not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "District"
        ]
      },
      "get": {
        "description": "เรียกดูตำบลทั้งหมด",
        "operationId": "DistrictController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operation successful.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Operation successful.",
                    "error": 0,
                    "data": [
                      {
                        "id": 1,
                        "name": "กรุงเทพ"
                      },
                      {
                        "id": 2,
                        "name": "ยโสธร"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": [
                      "Invalid input"
                    ],
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Resource not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "District"
        ]
      }
    },
    "/api/core/district/search": {
      "get": {
        "description": "เรียกดูตำบลตามเงื่อนไข",
        "operationId": "DistrictController_findByCondition",
        "parameters": [
          {
            "name": "textSearch",
            "required": false,
            "in": "query",
            "description": "Search text (comma-separated for multiple values)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "searchField",
            "required": false,
            "in": "query",
            "description": "Fields to search in (comma-separated for multiple fields)",
            "schema": {
              "example": "id,name",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "example": "1",
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "example": "10",
              "type": "string"
            }
          },
          {
            "name": "sortField",
            "required": false,
            "in": "query",
            "description": "Field to sort by",
            "schema": {
              "example": "id",
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sort order",
            "schema": {
              "example": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "amphure_id",
            "required": false,
            "in": "query",
            "description": "ค้นหาตาม id อำเภอ (amphure_id)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operation successful.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Operation successful.",
                    "error": 0,
                    "data": [
                      {
                        "id": 1,
                        "name": "กรุงเทพ"
                      },
                      {
                        "id": 2,
                        "name": "ยโสธร"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": [
                      "Invalid input"
                    ],
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Resource not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "District"
        ]
      }
    },
    "/api/core/district/{id}": {
      "get": {
        "description": "เรียกดูตำบลตาม id",
        "operationId": "DistrictController_findOne",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operation successful.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Operation successful.",
                    "error": 0,
                    "data": {
                      "name": "กรุงเทพ"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": [
                      "Invalid input"
                    ],
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Resource not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "District"
        ]
      },
      "patch": {
        "description": "แก้ไขข้อมูลตำบล",
        "operationId": "DistrictController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDistrictDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operation successful.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Operation successful.",
                    "error": 0,
                    "data": {
                      "id": 1,
                      "name": "ยโสธร2"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": [
                      "Invalid input"
                    ],
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "ต้องเข้าสู่ระบบก่อน",
                    "error": "Unauthorized",
                    "statusCode": 401
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "คุณไม่ได้รับอนุญาติให้กระทำสิ่งนี้",
                    "error": "Forbidden",
                    "statusCode": 401
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Resource not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "District"
        ]
      },
      "delete": {
        "description": "ลบตำบลตาม id",
        "operationId": "DistrictController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Operation successful.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Operation successful.",
                    "error": 0,
                    "data": {
                      "id": 1,
                      "name": "กรุงเทพ"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": [
                      "Invalid input"
                    ],
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "ต้องเข้าสู่ระบบก่อน",
                    "error": "Unauthorized",
                    "statusCode": 401
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "คุณไม่ได้รับอนุญาติให้กระทำสิ่งนี้",
                    "error": "Forbidden",
                    "statusCode": 401
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Resource not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "District"
        ]
      }
    },
    "/api/core/district/delete-many": {
      "post": {
        "description": "ลบตำบลหลายแถวด้วย id แบบ array",
        "operationId": "DistrictController_removeMany",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteManyDistrictDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operation successful.",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Operation successful.",
                    "error": 0,
                    "data": [
                      {
                        "id": 1,
                        "name": "กรุงเทพ"
                      },
                      {
                        "id": 2,
                        "name": "ยโสธร"
                      }
                    ]
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": [
                      "Invalid input"
                    ],
                    "error": "Bad Request",
                    "statusCode": 400
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "ต้องเข้าสู่ระบบก่อน",
                    "error": "Unauthorized",
                    "statusCode": 401
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "คุณไม่ได้รับอนุญาติให้กระทำสิ่งนี้",
                    "error": "Forbidden",
                    "statusCode": 401
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Resource not found",
                    "error": "Not Found",
                    "statusCode": 404
                  }
                }
              }
            }
          }
        },
        "summary": "",
        "tags": [
          "District"
        ]
      }
    }
  },
  "info": {
    "title": "API Title",
    "description": "API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    },
    {
      "url": "https://nakdev.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "x-version": {
        "type": "apiKey",
        "in": "header",
        "name": "x-version"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateDistrictDto": {
        "type": "object",
        "properties": {
          "amphure_id": {
            "type": "number",
            "description": "รหัสอำเภอจากตาราง amphure"
          },
          "name": {
            "type": "string",
            "description": "ชื่อตำบล"
          },
          "zip_code": {
            "type": "string",
            "description": "รหัสไปรษณีย์"
          }
        },
        "required": [
          "amphure_id",
          "name",
          "zip_code"
        ]
      },
...
      "UpdateDistrictDto": {
        "type": "object",
        "properties": {
          "amphure_id": {
            "type": "number",
            "description": "รหัสอำเภอจากตาราง amphure"
          },
          "name": {
            "type": "string",
            "description": "ชื่อตำบล"
          },
          "zip_code": {
            "type": "string",
            "description": "รหัสไปรษณีย์"
          },
          "id": {
            "type": "number",
            "description": "รหัสตำบล",
            "example": 2
          }
        },
        "required": [
          "id"
        ]
      },
      "DeleteManyDistrictDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "array",
            "description": "รหัสตำบล",
            "example": [
              1,
              2,
              5,
              8
            ]
          }
        },
        "required": [
          "id"
        ]
      }
    }
  }
}